+++ Libplanet.Explorer/GraphTypes/WorldStateType.cs
@@ -25,6 +25,12 @@ namespace Libplanet.Explorer.GraphTypes
                 resolve: context => context.Source.Legacy
             );
 
+            Field<NonNullGraphType<IntGraphType>>(
+                name: "version",
+                description: "The version of the bakcing data model.",
+                resolve: context => context.Source.Version
+            );
+
             Field<NonNullGraphType<AccountStateType>>(
                 name: "account",
                 description: